pyproject.toml
[project]
name = "le"
version = "0.1.0"
description = "LinkedIn Outreach and Business Process Optimization"
authors = [{ name = "AlexRook777" }]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "structlog>=24.0.0",
    "httpx>=0.27.0",
    "browser-use (>=0.11.9,<0.12.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "langchain-google-genai (>=4.2.0,<5.0.0)",
]

[tool.poetry]
packages = [{ include = "app" }]

[tool.poetry.group.dev.dependencies]
ruff = "*"
pytest = "*"
pytest-asyncio = "*"
pre-commit = "*"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "PERF", "FURB", "LOG", "RUF"]
ignore = ["D100", "D104"]

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
